updates test for test_empty_cf_output
This test depended on old cqlsh behavior.

Also, the test data from test_keyspace_init.cql used a timestamp that
would format correctly in some timezones, but not in others (e.g., in
Pacific time, the year was 1900, which formats correctly, but in
Eastern time, the year was 1899, which doesn't). I changed the year in
the timestamp so that the test won't fail based on the timezone of the
system under test.