fix cqlsh test plumbing
This fixes the cqlsh tests so they can run again. The role of the cqlsh
executable changed recently -- it became a shell script that calls
cqlsh.py. The plumbing for these tests expected cqlsh to be a python
script, which messed up the initial step of copying cqlsh to the test
directory.

There also were some import errors introduced by a restructuring in the
CQL parser, which this also fixes